reorganize code into cogs
