+++ code/cogs/__init__.py
@@ -1,0 +1,5 @@
+from cogs.setup import Setup
+from cogs.settings import Settings
+from cogs.help import Help
+from cogs.chat import Chat
+from cogs.manage_chat import ManageChat
